IPython >=1.1 and 2.x compatibility fixes
CAVEAT USER: The code is NOT actively developed on 1.1 and 2.x; it's
advisable to use 3.0 or later. We try to make it run on older versions on a
best-effort basis.

Also added a script into tests/ that can convert the overview notebook to an
older format that IPython <= 3.x can read. You can run this manually, and
check by hand if the code works on older IPythons.