Rename the wrapper script to ipython-eups
Instead of advising the user to place it ahead of the real 'ipython' on
their path, advise them to alias ipython to ipython-eups.  This solves the
problem where if the $PATH is modified, the real ipython may end up ahead of
the wrapper script (happens when ipython is managed by EUPS itself).